Fix minor bugs
When user inputs number like 0. (zero with decimal) and selects an operator, the screen will show the number like 0. with operator appended. This is not a valid number in maths although the calculator still gives the answer, it does not look right. So it was fixed by converting the string of number to actual number. Another bug was seen after fixing the previous bug, when the user does the previous operation and selects another number except zero or decimal, the number would be replaced at first which is what we want but that number would be replace again after selecting another number to type two digit numbers which is not what we want. That bug was fixed as well by setting new input value to false once the user clicks on any number.