Change variable name and fix minor bug
When new number was provided after mutliple continous operations, instead of using the new number, the number present on calculator screen would be used as second number. For example: User does addition of 5 + 3, which would give 8. User would then select any operator (let's use + for now) right after getting the result and click equals to without selecting second operand but the second operand would be replaced automatically with the answer, so the operation would be of 8 + 8 which would give 16 after clicking equals to sign. Then the user would click equals to sign again performing another operation of 16 + 8 (new answer) + (second operand/variable) which would give 24. Then the user would click new operator (let's take division) and provide new value 6 and click equals to sign. The operation should be 24/6 which should returh 4 but instead the operation takes place between 24/24 once the user clicks equals to sign. That is because the second operand/variable gets replaced by the new answer which was 24. This bug was fixed by simply making second operand/variable to not get replaced if new input is provided.