Fix UI glitch
Fix the glitch of number displaying after equals to sign is clicked and the decimal button is clicked right after. The screen should be cleared completely as the user will provide new input but, that's not happens. The input field will be cleared but the small text field will still have previous values until another number button is clicked. That was fixed by simply moving resetCalculator condition to the top. Another glitch was when user clicks on decimal button, the decimal would not be appended with zero. That was fixed as well.