Fix bug. Pass the value to assignOperand function when decimal button is clicked too.
When the user would click decimal button after providing any number (lets take 0.3) and providing any operator (lets take +), and for second operand, the user would provide 0. only. Since there are no digits after decimal, the number would simply be 0. So when user clicks on equals to sign the operation should be 0.3 + 0. However that is not the case, instead the operation takes place between 0.3 + 0.3. We know user did not provide 0.3 as the second operand, instead user provided 0. as second operand which should be converted to 0. But the second operand does not gets replaced with 0. Instead it is passed as an empty operand and when second operand is empty it will automatically gets replaced with first operand's value. That is because the function gets returned when user has already provided decimal to avoid multiple decimals so, the assignOperand function can never get the value of decimal as soon as it is clicked. That problem was fixed by simply assigning operand on decimal clicks as well.