Fix bug
After performing continuous calculations, if new operand is provided by clicking on any number button and equals to sign is pressed, the calculation will occur between new number and previous second operand number. It will give the proper answer but in the DOM small text field, it would look like calculation has occured between just recieved answer and previous operand. Example: when user provides new input, let's take 2, and hits equals to sign, the calculation will occur between 2 and previous second operand, let's take 4. The operator will be the last operator used, let's take multiplication. The answer recieved will be 8 which is correct but if you look at small screen where it shows the first operand and second operand with operator sign, it will look like 8 * 4 = 8 which is not true. The users might get confused if they check the operation and the answer. That's why the bug was fixed.