Prevent smallTextField value from being empty
This commit will fix the UI bug where smallTextField would be empty after providing new operand to do operation with just received answer. The bug would only appear after clicking equals to button to compute two numbers. Let's say we want to compute 5 + 3, and then click equals to button, we would get result 8. Then we select an operator right after getting the result and select new operand, let's take 5. Now the screen should be showing 8 + 5 but that's not what happens. Instead we would only see 5 and the "8 + " would not be seen which is inside smallTextField.

So the bug was fixed because it might cause confusion for users as they would not know what 5 is being added to until they have clicked on equals to button.