Fix UI problem: Display newly provided value instead of displaying answer after selecting operator
After getting the answer for any computation, let's take 5 + 3, the answer would be 8. Then if you want to perform another operation for different numbers, you would provide the first operand (let's take 2) and select an operator (let's take *). Then the screen should display newly provided number with the operator.

However, that's not what happens. Instead as soon as you provide the operator the screen will
display the answer instead of new number even though the computation will still happen between the new operands.

So, this bug was fixed because it might cause confusion for the users.