Create function for allClear button and initialise all the variables in calculator.js in one function to make it easier to reset all the values when needed
The main reason to set inital values for all varaibles in one function is so that the function can be called anytime we want to reset the values to their initial state.

For example, when allClear button is clicked all the values are reset so that user can start over again.

This also avoids DRY principle.