Move assignOperands outside of else block
The reason is so that it can assign operands everytime equals to button
is clicked.
This also fixes the issue where doing negate of a number, let's say 12
and then adding (or performing any operation) it to another number, let's
say 2, would give -10. Which is correct but when the user clicks on
equals to sign after that, it would not perform -10 + 2 immediatly. One
extra click was required to perform that task.